# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode:
    enum:
    - AUTH_001
    - AUTH_002
    - AUTH_003
    - AUTH_004
    - AUTH_005
    - AUTH_006
    - AUTH_007
    - USER_001
    - USER_002
    - USER_003
    - USER_004
    - VAL_001
    - VAL_002
    - VAL_003
    - VAL_004
    - DB_001
    - DB_002
    - DB_003
    - DB_004
    - DB_005
    - EXT_001
    - EXT_002
    - EXT_003
    - EXT_004
    - EMAIL_001
    - EMAIL_002
    - MQ_001
    - MQ_002
    - INTERNAL_001
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - CONFLICT
    - BAD_REQUEST
    - UNPROCESSABLE_ENTITY
    type: string
    x-enum-varnames:
    - CodeUnauthorized
    - CodeInvalidCredentials
    - CodeTokenExpired
    - CodeTokenInvalid
    - CodeInsufficientRights
    - CodeAccountLocked
    - CodeAccountNotActivated
    - CodeUserNotFound
    - CodeUserAlreadyExists
    - CodeInvalidUserData
    - CodeEmailNotVerified
    - CodeValidationFailed
    - CodeInvalidInput
    - CodeMissingField
    - CodeInvalidFormat
    - CodeDatabaseError
    - CodeRecordNotFound
    - CodeDuplicateEntry
    - CodeConstraintViolated
    - CodeTransactionFailed
    - CodeServiceUnavailable
    - CodeTimeout
    - CodeRateLimitExceeded
    - CodeCircuitBreakerOpen
    - CodeEmailSendFailed
    - CodeInvalidEmail
    - CodeMessagePublishFailed
    - CodeMessageConsumeFailed
    - CodeInternal
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeConflict
    - CodeBadRequest
    - CodeUnprocessableEntity
  github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail:
    properties:
      code:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode'
      detail:
        type: string
      instance:
        type: string
      meta:
        additionalProperties: true
        type: object
      status:
        type: integer
      time:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
...
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.Permission:
    properties:
      category:
        description: e.g., "user", "role", "dashboard"
        type: string
      created_at:
        type: string
      description:
        description: Human-readable description
        type: string
      id:
        type: string
      name:
        description: e.g., "user.create", "user.delete", "role.manage"
        type: string
      updated_at:
        type: string
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_mr-kaynak_go-core_internal_modules_identity_service.CreateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  github_com_mr-kaynak_go-core_internal_modules_identity_service.UpdateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    type: object
...
  internal_modules_identity_api.UserRoleRequest:
    properties:
      domain:
        type: string
      role:
        type: string
    required:
    - role
    type: object
...
host: localhost:3000
info:
  contact:
    name: Go-Core Team
    url: https://github.com/mr-kaynak/go-core
  description: Enterprise-grade Go boilerplate with REST API, gRPC services, and notification
    system
  license:
    name: MIT
  title: Go-Core API
  version: 1.0.0
paths:
...
  /policies/users/{user_id}/roles:
    delete:
      consumes:
      - application/json
      description: Remove a role from a user in a domain
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.UserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role removed
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Remove role from user
      tags:
      - Policies
    get:
      description: Get all roles assigned to a user
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - default: default
        description: Domain filter
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User roles
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Get user roles
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Assign a role to a user in a domain
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.UserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role added
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Add role to user
      tags:
      - Policies
  /roles:
    get:
      description: Get paginated list of roles
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: List all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role (admin only)
      parameters:
      - description: Role creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Create a new role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Delete a role by ID (admin only)
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Delete a role
      tags:
      - Roles
    get:
      description: Get role details by UUID
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Get a role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role details (admin only)
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated role
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Update a role
      tags:
      - Roles
  /roles/{id}/inherit/{parent_id}:
    delete:
      description: Remove role inheritance relationship
      parameters:
      - description: Child role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Parent role UUID
        in: path
        name: parent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hierarchy removed
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Remove role hierarchy
      tags:
      - Roles
    post:
      description: Set role inheritance (child inherits from parent)
      parameters:
      - description: Child role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Parent role UUID
        in: path
        name: parent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hierarchy set
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Set role hierarchy
      tags:
      - Roles
...
securityDefinitions:
  Bearer:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"